+++ packages/interactive_media_ads/android/src/main/kotlin/dev/flutter/packages/interactive_media_ads/InteractiveMediaAdsLibrary.g.kt
@@ -639,7 +639,13 @@ private class InteractiveMediaAdsLibraryPigeonProxyApiBaseCodec(
   override fun readValueOfType(type: Byte, buffer: ByteBuffer): Any? {
     return when (type) {
       128.toByte() -> {
-        return registrar.instanceManager.getInstance(readValue(buffer) as Long)
+        val identifier: Long = readValue(buffer) as Long
+        val instance: Any? = registrar.instanceManager.getInstance(identifier)
+        Log.e(
+          "${proxyApiCodecName(const KotlinOptions())}",
+          "Failed to find instance with identifier: \$identifier"
+        )
+        return instance
       }
       else -> super.readValueOfType(type, buffer)
     }